-Created multiple Use Cases -Created Util classes -Added City ViewModel -Cleaned up WeatherViewModel

+++ app/src/main/java/com/rektstudios/trueweather/data/local/RealmDaoImpl.kt
@@ -1,7 +1,7 @@
 package com.rektstudios.trueweather.data.local
 
-import com.rektstudios.trueweather.other.Constants.getDateEpochToday
-import com.rektstudios.trueweather.other.Constants.getTimeEpochNow
+import com.rektstudios.trueweather.domain.util.EpochUtil.getDateEpochToday
+import com.rektstudios.trueweather.domain.util.EpochUtil.getTimeEpochNow
 import io.realm.Realm
 import io.realm.Sort
 import io.realm.kotlin.toFlow
@@ -32,61 +32,57 @@ class RealmDaoImpl @Inject constructor(
         }
     }
 
-    override suspend fun addWeather(cityItem: CityItem, weatherHourItems: List<WeatherHourItem>) {
+    override suspend fun <T> addWeather(cityItem: CityItem, weather: T) {
         withContext(Dispatchers.IO) {
             realm.executeTransaction {
                 val city = getCityAsQuery(it,cityItem.cityName)
-                weatherHourItems.forEach {item ->
-                    city?.weatherEveryHour
-                        ?.where()
-                        ?.equalTo("timeEpoch",item.timeEpoch)
-                        ?.findAll()
-                        ?.deleteAllFromRealm()
+                when(weather){
+                    is WeatherHourItem -> {
+                        city?.weatherEveryHour
+                            ?.where()
+                            ?.equalTo("timeEpoch", weather.timeEpoch)
+                            ?.findAll()
+                            ?.deleteAllFromRealm()
+                        it.copyToRealm(weather)
+                        city?.weatherEveryHour?.add(weather)
+                    }
+                    is WeatherDayItem -> {
+                        city?.weatherEveryDay
+                            ?.where()
+                            ?.equalTo("dateEpoch", weather.dateEpoch)
+                            ?.findAll()
+                            ?.deleteAllFromRealm()
+                        it.copyToRealm(weather)
+                        city?.weatherEveryDay?.add(weather)
+                    }
                 }
-                it.copyToRealm(weatherHourItems)
-                city?.weatherEveryHour?.addAll(weatherHourItems)
             }
         }
     }
 
-    override suspend fun addWeather(cityItem: CityItem, weatherDayItem: WeatherDayItem) {
-        withContext(Dispatchers.IO) {
-            realm.executeTransaction {
-                val city = getCityAsQuery(it,cityItem.cityName)
-                city?.weatherEveryDay
-                    ?.where()
-                    ?.equalTo("dateEpoch", weatherDayItem.dateEpoch)
-                    ?.findAll()
-                    ?.deleteAllFromRealm()
-                it.copyToRealm(weatherDayItem)
-                city?.weatherEveryDay?.add(weatherDayItem)
-            }
-        }
-    }
-
-    override fun getCityList(): Flow<CityItem> {
-        return realm.where(CityItem::class.java)
+    override fun getCityList(): Flow<CityItem> =
+        realm.where(CityItem::class.java)
             .findAll()
             .asFlow()
             .flowOn(Dispatchers.IO)
-    }
 
-    private fun getCityAsQuery(iRealm: Realm, cityName: String = ""): CityItem?{
-        return iRealm.where(CityItem::class.java).equalTo("cityName", cityName).findFirst()
-    }
+    override fun getCity(city: String): Flow<CityItem?> = getCityAsQuery(realm,city).toFlow().flowOn(Dispatchers.IO)
+    private fun getCityAsQuery(iRealm: Realm, cityName: String = ""): CityItem? =
+        iRealm.where(CityItem::class.java).equalTo("cityName", cityName).findFirst()
 
-    override fun getCityWeatherCurrent(city: CityItem): Flow<WeatherHourItem?> {
-        return getCityAsQuery(realm,city.cityName)
+
+    override fun getCityWeatherCurrent(city: CityItem): Flow<WeatherHourItem?> =
+        getCityAsQuery(realm,city.cityName)
             ?.weatherEveryHour
             ?.where()
             ?.lessThan("timeEpoch", getTimeEpochNow())
             ?.sort("timeEpoch",Sort.DESCENDING)
             ?.findFirst()
             .toFlow()
-    }
+            .flowOn(Dispatchers.IO)
 
-    override fun getCityWeatherForecastInDays(city: CityItem): Flow<WeatherDayItem>? {
-        return getCityAsQuery(realm,city.cityName)
+    override fun getCityWeatherForecastInDays(city: CityItem): Flow<WeatherDayItem>? =
+        getCityAsQuery(realm,city.cityName)
             ?.weatherEveryDay
             ?.where()
             ?.greaterThanOrEqualTo("dateEpoch", getDateEpochToday())
@@ -94,10 +90,9 @@ class RealmDaoImpl @Inject constructor(
             ?.findAll()
             ?.asFlow()
             ?.flowOn(Dispatchers.IO)
-    }
 
-    override fun getCityWeatherForecastInHours(city: CityItem): Flow<WeatherHourItem>? {
-        return getCityAsQuery(realm,city.cityName)
+    override fun getCityWeatherForecastInHours(city: CityItem): Flow<WeatherHourItem>? =
+        getCityAsQuery(realm,city.cityName)
             ?.weatherEveryHour
             ?.where()
             ?.greaterThanOrEqualTo("timeEpoch", getDateEpochToday())
@@ -105,7 +100,5 @@ class RealmDaoImpl @Inject constructor(
             ?.findAll()
             ?.asFlow()
             ?.flowOn(Dispatchers.IO)
-
-    }
 
 }
